refactor: Modal 컴포넌트 구조 개선 및 합성 컴포넌트 패턴 적용
